+++ alicloud/import_alicloud_ots_table_test.go
@@ -6,7 +6,10 @@ import (
 	"github.com/hashicorp/terraform/helper/resource"
 )
 
-func TestAccAlicloudOtsTable_importBasic(t *testing.T) {
+// At present, OTS sdk does not support creating OTS instance and import test case does not support provider config,
+// so this test can not be run successfully.
+
+func SkipTestAccAlicloudOtsTable_importBasic(t *testing.T) {
 	resourceName := "alicloud_ots_table.basic"
 
 	resource.Test(t, resource.TestCase{